moved Elementary components to audio folder
