+++ audio/tsconfig.json
@@ -4,21 +4,6 @@
 		"*.md",
 		"*.vue"
 	],
-	"typedocOptions": {
-		"entryPoints": [
-			"../use/",
-			"./"
-		],
-		"out": "support/code/docs/",
-		"githubPages": false,
-		"entryPointStrategy": "expand",
-		"name": "Use-chromatone documentation",
-		"defaultCategory": "Chomatone",
-		"exclude": [
-			"**/aubio.js"
-		],
-		"readme": "use/README.md"
-	},
 	"compilerOptions": {
 		"baseUrl": ".",
 		"paths": {
